Nikolay/feature/msi upgrade

+++ desktop/build.ps1
@@ -1,4 +1,9 @@
-Remove-Item .\msi\dist -Recurse
+
+if(Test-Path '.\msi\dist') {
+    Remove-Item .\msi\dist -Recurse
+}
+
+.\Inc-Version.ps1
 
 pyinstaller --noconfirm --onefile --console `
 --distpath "msi/dist" `
@@ -15,3 +20,5 @@ pyinstaller --noconfirm --onefile --console `
 heartbeat.py
 
 copy .\config.json .\msi\dist\
+
+.\msi\BuildInstaller.ps1
